requirements.txt
# fastapi
# aiologger[aiofiles]
# pip-chill
# uvicorn
# gunicorn

fastapi==0.95.2
# aiologger[aiofiles]==0.6.1
pip-chill==1.0.3
uvicorn==0.22.0
gunicorn==20.1.0
pydantic==1.10.7
tavily-python==0.5.0
aiohttp
yandex-cloud-ml-sdk